Add "open with system default application" event to UI comm (#8588)
Addresses #5486 
Companion PR in ark: https://github.com/posit-dev/ark/pull/877

This PR creates a new event in the UI comm that says "open this thing
(file, presumably) with the default application according to the OS".

The immediate motivating example is to allow R's `browseURL()` to work
as expected in the Positron console.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- `browseURL()` in R now delegates to the operating system's default
opener for inputs that are not recognized as a web URL or an HTML file.

#### Bug Fixes

- N/A


### QA Notes

In R, use `browseURL()` on a local filepath. Exercise a couple of files
(meaning file types) and a folder. Here are examples taken from the "R
Code" column of the table at
https://github.com/posit-dev/positron/issues/5486#issuecomment-3081757100:

| R Code | Expectation in Positron R Console, with this PR + ark PR |
| -- | -- |
|`browseURL(tempdir())` | opens folder in OS's notion of file explorer
(e.g. Finder on macOS) |
| `temp_csv_file <- tempfile()` <br>`write.csv(iris,
temp_csv_file)` <br>`browseURL(temp_csv_file)` | opens the csv file in
OS's default application (for text files? on macOS, I'm getting
TextEdit) |
| `temp_csv_file_with_ext <- tempfile(fileext =
".csv")` <br>`write.csv(iris,
temp_csv_file_with_ext)` <br>`browseURL(temp_csv_file_with_ext)` | opens
the csv file in OS's default application for `.csv` (VS Code for me,
haha) |
| `temp_html_file <- tempfile()` <br>`writeLines("<p>Hi!</p>",
temp_html_file)` <br>`browseURL(temp_html_file)` | opens the html file
in OS's default application (for text files? on macOS, I'm getting
TextEdit) |
| `temp_html_file_with_ext <- tempfile(fileext =
".html")` <br>`writeLines("<p>Hi!</p>",
temp_html_file_with_ext)` <br>`browseURL(temp_html_file_with_ext)` |
opens `.html` file in default external browser |

When in doubt, the behaviour we seek is whatever happens in R in a plain
terminal or in RStudio Console.

We should definitely verify this on linux and Windows (I'm on macOS, so
feel pretty confident about that).

As for tests, I don't really know how we would do that 🤔 in this case.
The whole point is to open a file in *some other* application and
exactly what that is will vary by OS / user.

Signed-off-by: Jennifer (Jenny) Bryan <jenny.f.bryan@gmail.com>