data explorer: create "Copy as Code" modal (#8537)
This PR implements the actions and modal dialog for the `Copy as Code`
data explorer button.

A modal dialog was chosen over a modal popup since we cannot use a modal
popup in the editor action bar at this point in time. Notes from
@dhruvisompura:

> The data explorer UI doesn't have the action bar component we created
because we normally have it live in its own row below the VS Code editor
action bar (the one that every editor tab has). We don't want the data
explorer to have two action bars as that is confusing/annoying/bad ux
for the user! We are relying on the VS Code editor action bar for the
Data Explorer.
We have to contribute all of our buttons via a command for them to show
up in that editor action bar.
There currently is no mechanism to show a popup that is anchored to the
copy button via an action because commands have no way of knowing what
DOM element to anchor a piece of UI to. Commands are pretty primitive
and don't have UI context generally.

Note that this is just the first implementation! Things like the "Copy
to Clipboard" functionality, real code editor block, etc will come in a
follow up PR.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

Intended behavior is that the modal has a dropdown that is populated
with the name of a syntax, eg, `pandas`, + generated code. At this
point, the code generation piece is not implemented, so we won't have to
check the code is correct (yet).

On the Python side, a `pandas` dataframe will fill the dropdown with
`pandas`. A `polars` dataframe will fill the dropdown with `polars`. For
the R side, unless you have a locally built Ark, the button should be
greyed out since there are no syntaxes available.

---------

Signed-off-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>
Co-authored-by: Dhruvi Sompura <dhruvisompura@users.noreply.github.com>