Use table icon for non plaintext files in data explorer editor (#8701)
Addresses #8164

The data explorer editor did not provide any custom icons for the tab to
use, which meant that the default text file icon was used. We now show
the `table` icon if the editor tab is a data explorer input instance.
This is the same icon that we use in the variables pane for dataframe
objects.

The only open question as far as behavior is if we want to show a
different icon for data explorer editors backed by a plaintext resource.
Prior to this change, ".csv" files always displayed a custom csv icon in
the editor tab regardless if the file was viewed via the data explorer
or as a plaintext file. I am not sure if this was intentional and if we
want all editors backed by a csv resource to show the green csv icon.

The current behavior is such that every data explorer instance shows the
table icon. If a plaintext file is viewed in plaintext mode, the icon
will change to the text file icon (or the csv icon in the case of csv
files).



https://github.com/user-attachments/assets/83375d47-aa03-445f-be89-cb6d131afd1f


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Data Explorer editor tabs now use the `table` icon (#8164)

#### Bug Fixes

- N/A


### QA Notes

@:data-explorer @:layouts @:viewer @:editor @:variables @:web @:win