Data Explorer: Implement histogram as a single SVG path to avoid 1-pixel gaps depending on zoom level (#8662)
Addresses #8273. I had to seek Claude Code's help on this one as I am
not an expert in SVG rendering. By rendering the histogram as a single
SVG path, the UI scale level will not accidentally introduce 1-pixel
gaps as seen in #8273. The only extra complexity is that we need to
overlay transparent SVG rects for the histogram bar orange highlighting
to continue working.