Add `notebook.workingDirectory` setting (#8558)
<!-- Thank you for submitting a pull request.
If this is your first pull request you can find information about
contributing here:
  * https://github.com/posit-dev/positron/blob/main/CONTRIBUTING.md

We recommend synchronizing your branch with the latest changes in the
main branch by either pulling or rebasing.
-->

<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will
  automatically close the issue. If there are any details about your
  approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->
Addresses #7988. Adds a `notebook.workingDirectory` setting and hides
the `jupyter.notebookFileRoot` setting, which doesn't work in Positron.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Added `notebook.workingDirectory` setting, which can be used to set
the default working directory for notebooks (#7988)

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->
@:notebooks @:vscode-settings @:console @:extensions @:interpreter @:win

I added some unit tests and e2e tests!

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->
Some things to try:
- The `jupyter.notebookFileRoot` setting should no longer exist
- Open a workspace, and a new notebook in a subdirectory of that
workspace, with a cell containing `%pwd`, and see its value for
different values of the new setting (note you need to close/open the
notebook for the change to take effect):
  - (unset)
  - `${workspaceFolder}`
  - `${fileDirname}`
  - `${userHome}`
  - `/some/path/that/exists`
  - `/some/path/that/doesnt/exist`
- whatever else you want to try (variables are defined
[here](https://github.com/posit-dev/positron/blob/adcc385b17e3246c5439ea23c01eaf551b57403d/src/vs/workbench/services/configurationResolver/common/configurationResolver.ts#L77))
- The Console and Terminal working directories shouldn't change
- Filename completions should correspond to the correct working
directory
- This should work for R (`getwd()`) or Python (`os.getcwd()`)
notebooks, and for VSC or Positron notebooks