e2e: fix new folder flow conda test (#8789)
### Summary

This PR resolves a flake in the new folder flow test related to Conda
environment creation. The test would occasionally fail because the
`create-conda.py` script fell back to the defaults channel, despite CI
being configured to use `conda-forge` via Miniforge.

We could use either Miniforge or Miniconda, but for now we prefer
Miniforge due to recent changes to Miniconda that require Terms of
Service acceptance when using the defaults channel and this also aligns
with the container work done by @testlabauto.

### QA Notes

@:new-folder-flow @:web