e2e: support menuitemcheckbox in context menu (#8841)
### Summary
This PR includes several improvements to the `ContextMenu` POM:
* Added support for `checkboxmenuitem` in context menus (as need by
`Open In New Window`, etc), expanding beyond the previously supported
`menuitem` type.
* Updated menu item selection to support regular expressions, reducing
ambiguity when multiple items have similar labels.
* Refactored to remove the need for test-specific info, making the
implementation more self-contained and streamlined.

Updates to the `Plots` POM include:
* Added `setThePlotZoom()` method
* Added `openPlotIn()` method

### QA Notes

@:plots @:console @:win @:web

Note: the session-state failure on web is a flake and it's unrelated to
these chagnes. It's on my todo list. 😅