data explorer: `pandas` convert to code backend (#8716)
I've added a main `CodeConverter` class to be subclassed for each
package backend. Since we'll always know the data type at the time of
code conversion, there's no need for handling to dispatch a specific
backend. Instead, we'll directly call the appropriate converter (e.g.,
`PandasConverter`) when needed.

Inside the converter, we'll always follow these steps in order: 1)
handle row filters, 2) handle column filters (not yet implemented), and
3) handle sorts. Each step produces two lists of strings: one for
preprocessing and one for the method chain itself. The preprocessing
strings are concatenated into a single list, and the method chain is
combined into a single mega-string so the frontend can format it without
needing to handle escaped lines on the backend. The preprocessing +
method chain ends up as a resulting list to send to the front end.

Most of the code creation is using the same methods as in the
[`_eval_filter`
method](https://github.com/posit-dev/positron/blob/17ef3bb91e5a9f6bbf26366f195f007476fdff5a/extensions/positron-python/python_files/posit/positron/data_explorer.py#L1633)
that provides the Data Explorer itself with filter outputs, but there
are some differences:

- for date comparisons that include timezones: coerce the string value
to a `pd.Timestamp(value, tz=timezone)`^*
- for `RowFilterType.Between\NotBetween`: use the `.between(left,
right)` rather than the mask shown below

https://github.com/posit-dev/positron/blob/17ef3bb91e5a9f6bbf26366f195f007476fdff5a/extensions/positron-python/python_files/posit/positron/data_explorer.py#L1651-L1655
- for comparisons: rather than using `operator.operation_name(column,
value)`, use `df[df['column'] operation value]`, eg `df[df['column'] >
3]`

^*this is giving some slight inconsistencies 😩 pandas Timestamps are
inclusive of start dates whereas the Data Explorer is not. Looking into
options here! UPDATE: turns out this was a bug, see #8760.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #8530 Allows data explorer filters to be converted to code for
`pandas` data frames.

#### Bug Fixes

- N/A


### QA Notes

I've added a handful of unit tests here to check that the dataframe
generated from the "convert to code" is equivalent to an expected
output. We also have an e2e test that started running in #8693

You can do a quick overview by using the flights data in
`qa-example-content`. Unit tests cover all the operations, although not
for all types.

@:data-explorer

---------

Signed-off-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>