data explorer: convert to code UI update (#8829)
UI updates to the "Convert to Code" button!

- syntax highlighting for incoming code
- added a header for context on the dropdown of syntaxes
- changed "OK" button to "Copy Code" where code is copied to clipboard
and then modal is disposed

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #8696 Adds syntax highlighting to code converted from data explorer
filters.

#### Bug Fixes

- N/A


### QA Notes

with the `"dataExplorer.convertToCode": true` setting

```python
import pandas as pd

x = pd.DataFrame({"X":[1,2,3,4,5,6,7]})

%view x
```

and then click on the `Convert to Code` button. See syntax highlighting.

---------

Signed-off-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>
Co-authored-by: Dhruvi Sompura <dhruvi.sompura@posit.co>