Introduce Summary Panel Search Bar (#8721)
Addresses #8526, #8529

This feature is currently hidden behind the
`dataExplorer.summaryPanelEnhancements` setting. This setting is not
registered and will not show up in the settings UI. You can add
`dataExplorer.summaryPanelEnhancements: true` to your `settings.json`
file to enable the feature flag.

This PR introduces a basic implementation of search to the summary panel
in the Data Explorer that still uses the frontend search implementation.
The UI for the search in the summary panel is complete and should
unblock QA for writing e2e tests.

There will be follow up work to get the search backend piped through and
hooked up for the summary panel AND the data grid. This will be done
incrementally to prevent regressions which is why its not done in this
PR.

The rendering of the column profiles in the summary panel is not fully
working with search and will be fixed with the refactor once the search
API is introduced!

Note: The caching layer has not been optimized for the summary panel and
will be done when we add the backend search API. The bottleneck in the
current implementation is rendering the actual summary data (histograms,
etc). When viewing `flights.parquet` which is what we consider a dataset
representative of our users (~20 columns), there is latency due to us
re-fetching data and re-calculating these values.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

@:data-explorer @:duck-db @:web @:win