Add support for github refs when downloading Ark (#7645)
Adds support for specifying Ark versions as github refs in
`package.json`:

```ts
  "positron": {
    "binaryDependencies": {
      "ark": "posit-dev/ark@main"                  // Use the main branch
      "ark": "posit-dev/ark@experimental-feature"  // Use a feature branch
      "ark": "posit-dev/ark@a1b2c3d"               // Use a specific commit
      "ark": "posit-dev/ark@v0.1.183"              // Use a specific tag
    }
  }
```

The github revision is downloaded, built, and installed.

This allows CI-testing a branch in the Positron repo against a branch in
the Ark repo without having to release Ark first. Releasing just for the
purpose of testing is really not ideal as the feature might not be
working well yet and other developers might need to do further Ark
releases to make progress with their own work.

To prevent committing a dev ref that might have been forgotten in the
`package.json` file, a github action watches changes to this file and
checks that a release version is used.

<!-- Thank you for submitting a pull request.
If this is your first pull request you can find information about
contributing here:
  * https://github.com/posit-dev/positron/blob/main/CONTRIBUTING.md

We recommend synchronizing your branch with the latest changes in the
main branch by either pulling or rebasing.
-->

<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will
  automatically close the issue. If there are any details about your
  approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->