python: bump pet version (#8893)
related to https://github.com/posit-dev/positron-builds/issues/444


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Update to use latest Python Environment Tools for Python interpreter
discovery.


### QA Notes

green ci, and normal discovery! i clicked around and everything looks to
be acting as normal.