e2e: unskip convert to code test (#8951)
Flipping the switch for the e2e pandas conversion test!

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Run e2e test for converting pandas data explorer filters to code.

#### Bug Fixes

- N/A


### QA Notes

@:data-explorer should be green 💯