Summary panel data now hidden when collapsed (#9023)
Addresses #8989

This PR resolves the content overflow bug that caused the summary panel
data to be shown when the summary panel was collapsed.

The `SummaryRowActionBar` and `PositronDataGrid` were using different
approaches to be hidden when the summary panel was collapsed. The
`PositronDataGrid` is always rendered in the DOM, and hidden via CSS
width styling `width: 0` and `overflow: hidden` to prevent content from
spilling out when the container width is 0.

Rather than conditionally showing/hiding the `SummaryRowActionBar` when
the summary panel is collapsed, we are now using CSS to hide the content
like the `PositronDataGrid`. The `SummaryRowActionBar`. This change
prevents the component from un-mounting when the summary panel is
collapsed which means the user's last used state is now preserved until
the Data Explorer is closed.

Both components now use the same CSS visibility solution.

### Release Notes 

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

@:data-explorer