Assistant: always display `Working...` indicator at bottom of response while in progress (#8900)
### Summary

- addresses https://github.com/posit-dev/positron/issues/7740

Before
- Ask Mode: `Working...` chat progress indicator was displayed inline in
the chat avatar row
- Edit/Agent mode: `Working...` chat progress indicator was displayed
below the chat avatar row
- When markdown content is being streamed, the `Working...` chat
progress indicator is NOT shown if the response is still in progress

After
- Ask/Edit/Agent modes: `Working...` chat progress indicator was
displayed at the bottom of the response as long as the response is still
in progress
- Note that this makes the chat look a bit more jumpy/in motion, so we
may want to create a new element like the chat total token count, but we
may not be able to access the individual chat item state as easily in
that part of the code

#### Demo


https://github.com/user-attachments/assets/56319b54-8414-4248-a42f-30284b58623d

### Release Notes

#### New Features

- Assistant: display `Working...` indicator at bottom of response while
response is in progress (#7740)

### QA Notes

@:assistant

Although the indicator is more apparent now, I was able to confirm that
the indicator disappears before the assistant response is done streaming
in the UI.

To be extra sure that the response is done, it'd be best to wait for the
element matching the selector
`.chat-most-recent-response.chat-response-loading` to not be visible.
The most recent response has the `chat-response-loading` class applied
while the chat is still in progress. The `Working...` indicator and
send/stop button both switch slightly earlier than the
`chat-response-loading` class.