data explorer: add row filtering context key (#8975)
Adding a context key, similar to the `isColumnSorting`, but instead
we'll track if there's filtering. I'm just using the backend state to
see if any filters are present, since it syncs as filters get
applied/removed.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #8719 Adds a context key that tracks if row filters are applied.

#### Bug Fixes

- N/A


### QA Notes

The `Convert to Code` button should only be available in `pandas`
DataFrames when either filtering or sorting is applied. Here's a simple
df to copy/paste for convenience!


```python
import pandas as pd

x = pd.DataFrame({'x':[1,2,3,4,5]})
%view x
```