Update 1Password action to v3 (#9058)
The 1Password GH Action now supports windows as well.

Currently this just sets an anthropic key for assistant testsing, but can now be utilized for any secret.

The assistant test that uses it is skipped for now, but this is needed when it's enabled soon.

### QA Notes
In testing, I unskipped the test and ran it in CI. You can see it passed in https://github.com/posit-dev/positron/actions/runs/16993754640/job/48179307637

It's marked failed due to the S3 upload of playwright log failure.