bump ark to 0.1.202 (#9100)
Bringing in
- https://github.com/posit-dev/ark/pull/898
- https://github.com/posit-dev/ark/pull/893
- https://github.com/posit-dev/ark/pull/896
- https://github.com/posit-dev/ark/pull/891
- https://github.com/posit-dev/ark/pull/889


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- https://github.com/posit-dev/positron/issues/8805 Implement R search
schema RPC calls for data explorer.
- Implement RPC calls for sorting columns by type.

#### Bug Fixes

- https://github.com/posit-dev/positron/issues/8794 Update command name
for `rstudioapi` session restart.
- Hide Convert to Code button for not implemented data types.
- Fix warnings for hidden lifetimes in ARK.


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->