Debugging notebooks (#8983)
This PR adds support for debugging notebooks, addressing #3906.

This is a pretty big change so I expect there to be rough edges. Where
possible I'd like to defer fixes to be in separate PRs so we can get the
foundation in.

### Release Notes

#### New Features

- You can now debug Python Jupyter notebooks.

#### Bug Fixes

- N/A

### QA Notes

* Open a Python notebook and try the Debug Cell command.
* Try out edge cases like interrupting a cell while debugging,
restarting the runtime, debugging function calls across cells, etc.