data explorer: remove convert-to-code feature flag (#9110)
Removes feature flag from implemented convert to code backends.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

- Convert to code button SHOULD show up for pandas dataframes
- Convert to code button SHOULD NOT show up for polars dataframes, any R
dataframes, csvs, parquets, etc