python: make PET default locator (#9111)
Changes from "js" locator to the PET. 🎉 

This PR also removes CI runs for upstream smoke tests. This was brought
on due to the fact that smoke tests do not run from the expected
`positron/extensions/positron-python` directory, so the `pet` binary
would need to be redownloaded and inserted into the temporary testing
path. We have disabled all behavior that is being tested in the smoke
tests besides: checking if the LSP is on, running a Python file in the
terminal, and checking if Smart Send is on. These are all features that
are covered in Positron's own e2e tests. Since the smoke tests are not
adding additional coverage or information to our testing suite, we
decided to remove them completely.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #6582 Use Rust-based Python Environment Tools locator as default.

#### Bug Fixes

- N/A


### QA Notes

passes CI and if you poke around in different workspaces, Python should
be discovered. I don't think we need to add "native" to the e2e settings
anymore.