Send pre-renderings with plot updates (#8942)
Addresses https://github.com/posit-dev/positron/issues/7679
Addresses https://github.com/posit-dev/positron/issues/8941

Since https://github.com/posit-dev/positron/pull/7247 we send
pre-renderings of plots in the `comm_open` notification that opens the
plot, so that it can be immediately displayed in the plot pane. This
isn't the case for plot _updates_ which currently require the frontend
to request the rendering, resulting in delays and missing updates. To
fix this, the `update` notification gains an optional `pre_render`
parameter that the frontend can use to display the update immediately.
If the rendering settings are outdated and do not match, we still
request a rerender as before.

In addition, this PR fixes an issue in the comm generation where
importing from the backend in the frontend of a comm (or vice versa)
resulted in a redundant import.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Plots created in steps with prompts in between (e.g. `demo(graphics)`
are now immediately displayed in full (#7679).
- Plots created in steps are now consistently rendered with expected
rendering settings (#8941).


### QA Notes

Run an expression that creates a plot in two steps:

```r
dat <- mtcars
pr <- princomp(dat)
biplot(pr)
```

With release Positron there is a delay between the two steps:


https://github.com/user-attachments/assets/65fb5ea4-659c-45a0-b054-e07db260731c

With the fix, it's much snappier:


https://github.com/user-attachments/assets/ea5dc4f2-3600-4f46-86bc-69f5b4c8beb7

If you create the plot a second time, with release Positron the second
plot uses incorrect rendering settings:


https://github.com/user-attachments/assets/16dd9652-b521-4257-8474-61fa48c1b7c4

With the fix, the second plot is rerendered with the right settings
immediately:


https://github.com/user-attachments/assets/3ec3abf3-2921-497c-9303-099f342acb5c

More important fix: With an expression that prompts the user for inputs
between multi-step plots:

```r
demo(graphics)
```

With release Positron, you only get updates after the whole command
returns


https://github.com/user-attachments/assets/c593e948-cba9-4d6c-8966-fc8d918f3698

With the fix, the plot is fully rendered immediately:


https://github.com/user-attachments/assets/ea96eab4-f2ef-4a99-93a1-3d1b03c8d50b